Add FLAVOR build option to select pins to use.
Use "make PROBE_HOST=nsec17 FLAVOR=crossdebug" to build the firmware that uses
the exposed pins to debug another badge. Setting FLAVOR to anything else will
build the firmware to debug the onboard nRF51.

This should really be a runtime option to allow a single firmware to debug both
the internal nRF51 and an external stm32, but it's a compile time option for now
due to a lack of time before the event. This can be added later anyway. Feel
free to implement it :)