Don't define BLACKMAGIC
The usbusart handler (for host -> target communication) contains this
check to see if the USART device is enabled:

    #if defined(BLACKMAGIC)
    	/* Don't bother if uart is disabled.
    	 * This will be the case on mini while we're being debugged.
    	 */
    	if(!(RCC_APB2ENR & RCC_APB2ENR_USART1EN))
    		return;
    #endif

The register and bit used assume that we are using USART1, which is not
our case.  To avoid it, we can avoid defining the BLACKMAGIC macro.
This seems to be the only place where this macro has an effect.  And in
our case, we expect the USART to be always on.

With this change, and additional changes on the nRF52, I am able to send
bytes from the development machine to the nRF52.