usbusart: Clear error flags when USART receives an error
When the USART encounters an error (we currently check for ORE (overflow
error) and FE (framing error)), we don't clear the flags, which means
they get set forever.

The datasheet says clearly, about the ORE bit:

    This bit is set by hardware when the data currently being received in the shift
    register is ready to be transferred into the RDR register while RXNE=1. It is
    cleared by a software, writing 1 to the ORECF, in the USART_ICR register.

So once it is set, the USBUSART_ISR function will always return early, and
never transmit the received bytes on the USB again.